pdf-agent/requirements.txt
pymupdf==1.25.5
PyPDF2==3.0.1
python-dotenv==1.1.0
tiktoken==0.7.0
pyyaml==6.0.2

# For our components
faiss-cpu
sentence-transformers
langchain
requests
